Create 3009_Chasing_Game.c
Base game thread,
The game is played on the alphanumerical terminal and 4 directional buttons
are used to control the movements. Two objects are displayed: the moving target
(rabbit) and the chaser (fox). Both are represented as a single character and moved
at the speed of 1 position every 0.5 seconds. The rabbit makes random turns and
never leaves the screen. The turns of the fox are controlled by the player.